Update Twitter oauth module
Update refreshToken message because it was confusing whether it was
coming from the revoke or refresh method. It was more confusing because
I haven't touched on this module for so long.

It seems like Twitter invalidate API works properly now. Because it works
I now know that it returns something, and return the response of
invalidate request.