Add isRefreshTokenError & update Github
`isRefreshTokenError` can tell if the error is due to refreshing with
tokens that are invalid possibly due to revoking on the service's
website.

Apparently, all of the services that actually allow refresh (all oauth2
except Github) returns 400 error as stated in the oauth2 spec.

Github tokens API has been updated:

  - https://developer.github.com/changes/2020-02-14-deprecating-oauth-app-endpoint/